deploy/requirements.txt
autopep8==1.6.0
bleach==4.1.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
cryptography==36.0.1
docutils==0.18.1
flake8==4.0.1
google-auth==1.28.1
idna==2.10
importlib-metadata==4.2.0
importlib-resources==5.4.0
jeepney==0.7.1
Jinja2==2.11.3
keyring==23.4.1
kubernetes==12.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
numpy==1.13.1
oauthlib==3.1.0
openshift==0.12.0
packaging==21.3
pkginfo==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.1
python-string-utils==1.0.0
PyYAML==5.4.1
readme-renderer==32.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scipy==0.19.1
SecretStorage==3.3.1
simple-http-server==0.12.4
six==1.15.0
toml==0.10.2
tqdm==4.63.0
twine==3.8.0
typing_extensions==4.1.1
urllib3==1.26.4
virtualenv==15.1.0
webencodings==0.5.1
websocket-client==0.58.0
zipp==3.6.0
